Use %lu with (unsigned long) casts for uid_t and gid_t
POSIX only guarantees uid_t/gid_t are integer types, not specifically
unsigned int. Cast to (unsigned long) with %lu to avoid truncation on
platforms where these types are wider than unsigned int. Also updates
the two existing (unsigned)/%u usages for consistency.

+++ user.c
@@ -78,14 +78,14 @@ int setup_groups(uid_t uid)
     errno = 0;
     ps = getpwuid(uid);
     if (ps == NULL) {
-        error("Cannot get passwd info for uid %u: %s", (unsigned)uid, strerror(errno));
+        error("Cannot get passwd info for uid %lu: %s", (unsigned long)uid, strerror(errno));
         exit(ROOT_SYSTEM_ERROR);
     }
 
     errno = 0;
     result = setgid(ps->pw_gid);
     if (result == -1) {
-        error("Cannot setgid %u: %s", (unsigned)ps->pw_gid, strerror(errno));
+        error("Cannot setgid %lu: %s", (unsigned long)ps->pw_gid, strerror(errno));
         exit(ROOT_SYSTEM_ERROR);
     }
 
@@ -112,7 +112,7 @@ int set_home_dir(uid_t uid)
     errno = 0;
     ps = getpwuid(uid);
     if (ps == NULL) {
-        error("Cannot get passwd info for uid %u: %s", (unsigned)uid, strerror(errno));
+        error("Cannot get passwd info for uid %lu: %s", (unsigned long)uid, strerror(errno));
         exit(ROOT_SYSTEM_ERROR);
     }
 
@@ -143,7 +143,7 @@ int become_user(uid_t uid)
      */
     errno = 0;
     if (setuid(ROOT_UID) == -1) {
-        error("Cannot setuid %u: %s", (unsigned)ROOT_UID, strerror(errno));
+        error("Cannot setuid %lu: %s", (unsigned long)ROOT_UID, strerror(errno));
         return 0;
     }
     return 1;
